svn:revision:59386
svn:branch:nurbs
svn:account:indianlarry

+++ src/other/stepcode/src/cllazyfile/sectionReader.cc
@@ -83,7 +83,7 @@ std::streampos sectionReader::findNormalString( const std::string & str, bool se
 const char * sectionReader::getDelimitedKeyword( const char * delimiters ) {
     static std::string str;
     char c;
-    str.assign( 0, 0 ); //clear() frees the memory
+    str.assign( (size_t) 0, 0 ); //clear() frees the memory
     str.reserve( 100 );
     skipWS();
     while( c = _file.get(), _file.good() ) {
